moved login button events from api.js to ui_login.js
