+++ src/api/api.js
@@ -1,40 +1,13 @@
 import * as globals from "../util/globals.js";
 import { getEditor } from "../input/editor.js";
-import { showPopUpNotification } from "../ui/ui.js";
 
 let stored_username;
-let loginButton = document.getElementById("login-button");
-let logoutButton = document.getElementById("logout-button");
 const url = 'http://localhost:3000/api/';
 
-// eventlisteners should probably eventually be moved elsewhere
-loginButton.addEventListener("click", () => {
-    login(url)
-        .then(data => {
-            if (data.token !== undefined) {
-                localStorage.setItem("token", data.token);
-                localStorage.setItem("username", stored_username);
-                window.location.reload();
-            } else {
-                showPopUpNotification("login-failed");
-            }
-        })
-        .catch(error => {
-            console.error('Error:', error);
-        });
-});
 
-logoutButton.addEventListener("click", () => {
-    logout(url)
-        .then(function() {
-            localStorage.removeItem("token");
-            localStorage.removeItem("username");
-            window.location.reload();
-        })
-        .catch(error => {
-            console.error('Error:', error);
-        });
-});
+export function getStoredUsername() {
+    return stored_username;
+}
 
 /**
  * a function used to create api URLs from a base URL and an object
@@ -88,12 +61,9 @@ async function sendGetRequest(url) {
 /**
  * a function used to login to the game
  * @returns a JSON object containing a token, for example: { token: "45h239809sdfy8s32g3h23u" }
- * the token is stored in localstorage and is required for all other api calls
+ * the token is stored in localstorage and is required for all other api calls.
  */
-export async function login() {
-    const user = document.getElementById("username").value;
-    const pass = document.getElementById("password").value;
-
+export async function login(user, pass) {
     if (user === "") {
         showPopUpNotification("login-failed");
         return
